Add clean-orders command to remove spool IDs from wrong locations
When locations are updated through the UI in the locations tab,
the spool IDs of the moved spools are left in the source location. The
UI just shows the spools in the location that the spool is associated
to, and uses the locations_spoolorders values to keep track of each
spool's position within a location. This clean-orders command
removes spool IDs from locations in the locations_spoolorders
setting where the spool id is in a location that doesn't match what
the position the spool is associated to in the spool table.